feat(seguimiento): update user ID type to Integer in service and repository methods and simplify dashboard retrieval by removing user ID parameter and adding authorization

+++ src/main/java/com/sistemas_mangager_be/edu_virtual_ufps/modulo_seguimiento/controllers/ColoquioController.java
@@ -32,7 +32,7 @@ public class ColoquioController {
     }
 
     @GetMapping("/usuario/{usuarioId}")
-    public List<ColoquioDto> obtenerColoquiosPorUsuarioId(@PathVariable Long usuarioId) {
+    public List<ColoquioDto> obtenerColoquiosPorUsuarioId(@PathVariable Integer usuarioId) {
         return coloquioService.obtenerColoquiosPorUsuarioId(usuarioId);
     }
